requirements.txt
streamlit
numpy
torch
matplotlib
Pillow
ultralytics